# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/customer": {
      "get": {
        "tags": [
          "customer"
        ],
        "summary": "List all customers",
        "operationId": "getCustomers",
        "responses": {
          "200": {
            "description": "Fetched customer successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "408": {
            "description": "Invalid id supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "Create customer",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created customer",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input was provided",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          }
        }
      }
    },
    "/customer/{id}": {
      "get": {
        "tags": [
          "customer"
        ],
        "summary": "Get customer by id",
        "operationId": "getCustomerById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer Id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched language by id successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "customer"
        ],
        "summary": "Delete Customer by id",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer Id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Delete Customer by id"
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Customer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          }
        }
      },
      "ReturnModel": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "body": {
            "type": "object"
          },
          "successful": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}